Refactor library to allow us to add more modules underneath node_base
